requirements.txt
anthropic>=0.42.0
google-genai>=1.0.0
edge-tts>=6.1.0
fish-audio-sdk>=0.1.0
eval-type-backport>=0.2.0
faster-whisper>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
